+++ src/main/java/net/laboulangerie/laboulangeriemmo/abilities/thehunter/Dodging.java
@@ -6,8 +6,8 @@ import net.minecraft.world.entity.player.Player;
 import org.bukkit.event.Event;
 import org.bukkit.util.Vector;
 
-import net.laboulangerie.laboulangeriemmo.abilities.AbilityExecutor;
-import net.laboulangerie.laboulangeriemmo.abilities.AbilityTrigger;
+import net.laboulangerie.laboulangeriemmo.api.ability.AbilityExecutor;
+import net.laboulangerie.laboulangeriemmo.api.ability.AbilityTrigger;
 import net.laboulangerie.laboulangeriemmo.core.combo.ComboKey;
 import net.laboulangerie.laboulangeriemmo.core.combo.KeyStreak;
 import net.laboulangerie.laboulangeriemmo.events.ComboCompletedEvent;
