+++ src/main/java/net/laboulangerie/laboulangeriemmo/events/MmoPlayerUseAbilityEvent.java
@@ -7,7 +7,7 @@ import org.bukkit.event.HandlerList;
 import org.jetbrains.annotations.NotNull;
 
 import net.laboulangerie.laboulangeriemmo.abilities.Abilities;
-import net.laboulangerie.laboulangeriemmo.abilities.AbilityTrigger;
+import net.laboulangerie.laboulangeriemmo.api.ability.AbilityTrigger;
 import net.laboulangerie.laboulangeriemmo.api.player.MmoPlayer;
 
 /**
